requirements.txt
Flask
numpy
mysql-connector-python
joblib
scikit-learn
